+++ MameBake3D/FpsSprite.cpp
@@ -132,7 +132,7 @@ int CFpsSprite::SetParams(int srcmainwidth, int srcmainheight)
 	int spashift = 1;
 
 	//m_pointR[0].x = 175;
-	m_point[0].x = 180 - spawidth - spashift;
+	m_point[0].x = 180 - (int)spawidth - spashift;
 	m_point[0].y = 10;
 
 	m_point[1].x = m_point[0].x + (int)spawidth * 2 + spashift;
